후기 추가,수정 파일 취합 및 주석 추가
