+++ AnimalHospitalProject/src/main/resources/static/mapper/postscriptMapper.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" 
                         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="kr.human.anihospital.mapper.PostscriptSaveMapper">
+<mapper namespace="kr.human.anihospital.mapper.PostscriptMapper">
 	<insert id="insertPostscript"
 		parameterType="kr.human.anihospital.vo.PostscriptVO"> 
 			insert into Postscript(
@@ -24,4 +24,14 @@
 				now()
 			)
 	</insert>
+	<update id="updatePostscript" parameterType="kr.human.anihospital.vo.PostscriptVO">
+		<![CDATA[
+		    update  postscript SET
+		    		postscript_content = #{postscriptContent},
+		    		postscript_satisfaction = #{postscriptSatisfactionrating},
+		    		postscript_update_date = now()
+		    where
+		    		seq_postscript = #{seqPostscript}
+	   ]]>
+	</update>
 </mapper>
