--- AnimalHospitalProject/src/main/resources/static/mapper/postscriptUpdateMapper.xml
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" 
-                        "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="kr.human.anihospital.mapper.PostscriptUpdateMapper">
-	<update id="updatePostscript" parameterType="kr.human.anihospital.vo.PostscriptVO">
-		<![CDATA[
-		    update postscript SET
-		    		postscript_content = #{postscriptContent},
-		    		postscript_satisfaction = #{postscriptSatisfactionrating},
-		    		postscript_update_date = now()
-		    where
-		    	seq_postscript = #{seqPostscript}
-	   ]]>
-	</update>
-</mapper>
